Save non working version of Ant stubs with Roles added
TOFIX :
- We want the concrete Roles (= States) to be singletons, but each Ant
should have their own roles when the decide their actions : we need to
make sure that Ant::_current_role is a member pointer to a Singleton
RoleAnt
- The constructor of the RoleAnt subclasses is not protected so it
breaks the Singleton pattern. Currently it is this way because we want
to fill the RoleAnt::_registry with global instances of RoleAnt
subclasses. But it doesn't work ("explorer" throws a RoleNotFound in
main) so this needs a total rework probably
- Also, Entity class need a setter for acceleration, since LivingEntity
states set their context's acceleration
- This means that once VelocityStrategies are in place, we will need
helper methods in Entity to set a desired velocity/force.