Refactor _vision out of Ant hands
Having to rewrite versions of is_visible and boundingRect in Ant meant
that something was wrong in the way LivingEntity::_vision is handled.

Now, LivingEntity::_vision is a QSharedPointer<QPolygonF> with a setter,
so subclasses of LivingEntity and their users have the responsibility
of using this setter to keep _vision updated and therefore no
specialization of LivingEntity::is_visible() or boundingRect() is ever
needed.

Fixes #8