Fix an issue where Ant would go to infinity with only Cohesion
The issue was quite long to solve because the main issue was the way
data was transferred from one Decorator to the next.

The actual answer was to never trust Decorators to handle data and just
always use the context pointer to modify the variable.

NOTE : current context is an Ant* but for extensibility's sake, it would
probably be better to have Role* as a context, and then make Role a
superclass of RoleAnt and then pass RoleAnt as context instead of the
actual Ant. RoleAnt is then responsible for forwarding a new
RoleAnt::context() call to the good Ant* so the current behaviour of
MovementStrategy still works