Add a toggle for showing the vision of entities in the window
This most notably needed to pull all vision logic up from LivingEntity
to Base Entity class, so World can more easily call the toggle on all
its items without the pain of having to check super_type_string() and
cast.

Since base Entity class does not care in its implementation about
vision, _vision becomes basically a dummy field for InertEntites, and if
space is ever a concern it can be changed to a pointer anyway. Having a
pointer also means that only the QPolygonF is allocated instead of the
whole Entity object, but this needs to be check (What does a QPolygonF
object look like in memory ? Is it just a pointer to some memory
allocated elsewhere or is the whole structure living in the object
memory space ?)