World : Implement wrap_around so it fixes positions
The tests only test that the World::fix_position_for_wrap_around method
is working well enough, to correct the position and display of an entity
which went beyond World boundaries and should appear on the other side.

The feature is not over with since :
- the tests are failing because of floating point comparison (hence the
commit to show the code and get help online)
- the LivingEntity::is_visible (which should also go to base class by
the way) method is not taking the wrap_around into account : Enties
don't see in front of them like the world is actually of finite size.
This will be fixed later