Entity : Move boundingRect() and update_neighbourhood() up
Since the vision handling is moved up from LivingEntity to Entity, a few
overrides are not necessary anymore. We can directly check for the
vision in Entity::update_neighbourhood and correct the boundingRectangle
(useful for lazy canvas update) in Entity::boundingRect too.

We only had to add a check for _vision.isNull() since it is indeed null
for InertEntity